requirements.txt
openai-agents
streamlit
pypdf
python-dotenv
litellm
google-generativeai
